Basic annotations - add
Basic annotations - add

+++ 0x00-python_variable_annotations/0-add.py
@@ -2,7 +2,7 @@
 """ a type-annotated function add """
 
 
-def add (a: int, b: float) -> float:
+def add (a: float, b: float) -> float:
     """[type-annotated add function]
 
     Args:
